:sparkles: Add a PUT endpoint.
Presently the endpoint does nothing but canonicalize the annotations.

Canonicalization ensures that the same set of (Predicate,ConceptId) values will be considered to represent the same set of annotations (with hash calculations returning the same value), irrespective of the order in which they are presented by the client, and irrespective of additional mutable fields such as prefLabel.

For canonicalization, annotations are sorted lexicographically, first by Predicate, and then by ConceptId within a set of values for the same Predicate. (Strictly the specification doesn't need to be known externally, as only this service will ever calculate it. It just needs to be reproducible.)